+++ src/oeis/oeis_program.cpp
@@ -76,10 +76,10 @@ size_t OeisProgram::getNumMinimizationTerms(const Program &p) {
   return getNumRequiredTerms(p) * 2;  // magic number
 }
 
-std::vector<bool> OeisProgram::collectLatestProgramIds(
-    size_t max_commits, size_t max_added_programs,
-    size_t max_modified_programs) {
-  std::vector<bool> latest_program_ids;
+UIDSet OeisProgram::collectLatestProgramIds(size_t max_commits,
+                                            size_t max_added_programs,
+                                            size_t max_modified_programs) {
+  UIDSet latest_program_ids;
   auto progs_dir = Setup::getProgramsHome();
   if (!isDir(progs_dir + ".git")) {
     Log::get().warn(
@@ -91,7 +91,6 @@ std::vector<bool> OeisProgram::collectLatestProgramIds(
     Log::get().warn("Cannot read programs commit history");
     return latest_program_ids;
   }
-  std::set<int64_t> ids;
   size_t num_added_ids = 0, num_modified_ids = 0;
   for (const auto &commit : commits) {
     if (num_added_ids >= max_added_programs &&
@@ -109,12 +108,12 @@ std::vector<bool> OeisProgram::collectLatestProgramIds(
           if (isFile(ProgramUtil::getProgramPath(uid))) {
             if (status == "A" && num_added_ids < max_added_programs) {
               Log::get().debug("Added program for " + uid.string());
-              ids.insert(uid.number());
+              latest_program_ids.insert(uid);
               num_added_ids++;
             } else if (status == "M" &&
                        num_modified_ids < max_modified_programs) {
               Log::get().debug("Modified program for " + uid.string());
-              ids.insert(uid.number());
+              latest_program_ids.insert(uid);
               num_modified_ids++;
             }
           }
@@ -123,14 +122,6 @@ std::vector<bool> OeisProgram::collectLatestProgramIds(
         }
       }
     }
-  }
-  for (auto id : ids) {
-    if (id >= static_cast<int64_t>(latest_program_ids.size())) {
-      const size_t new_size =
-          std::max<size_t>(id + 1, 2 * latest_program_ids.size());
-      latest_program_ids.resize(new_size);
-    }
-    latest_program_ids[id] = true;
   }
   if (latest_program_ids.empty()) {
     Log::get().warn("Cannot read programs commit history");